Add files via upload
This text file can be copied into SQL.
The code answers Questions 1-6 from the Henry Meds SQL Assessment V2.
NOTE:  Question #6 mentioned a "Product Name" field.  This is not a field in the 3 table schemas mentioned in the assessment. I left the field as NULL for now.  The inclusion of a "Products" table would likely have a Product Name field that we could add to Question #6 to fulfill that request.